+++ Kernel/Modules/AgentTicketBulk.pm
@@ -98,7 +98,8 @@ sub Run {
             Valid => 1
         );
 
-        # Put only possible rw agents to owner list.
+        # Put only possible 'owner' and 'rw' agents to owner list.
+        my %OwnerGroupsMembers;
         if ( !$ConfigObject->Get('Ticket::ChangeOwnerToEveryone') ) {
             my %AllGroupsMembersNew;
             my @QueueIDs;
@@ -124,21 +125,24 @@ sub Run {
                 my $GroupID = $QueueObject->GetQueueGroupID( QueueID => $QueueID );
                 my %GroupMember = $GroupObject->PermissionGroupGet(
                     GroupID => $GroupID,
-                    Type    => 'rw',
+                    Type    => 'owner',
                 );
                 USER_ID:
                 for my $UserID ( sort keys %GroupMember ) {
                     next USER_ID if !$AllGroupsMembers{$UserID};
                     $AllGroupsMembersNew{$UserID} = $AllGroupsMembers{$UserID};
                 }
-                %AllGroupsMembers = %AllGroupsMembersNew;
+                %OwnerGroupsMembers = %AllGroupsMembersNew;
             }
+        }
+        else {
+            %OwnerGroupsMembers = %AllGroupsMembers;
         }
 
         my @JSONData = (
             {
                 Name         => 'OwnerID',
-                Data         => \%AllGroupsMembers,
+                Data         => \%OwnerGroupsMembers,
                 PossibleNone => 1,
             }
         );
@@ -148,9 +152,51 @@ sub Run {
             && $ConfigObject->Get("Ticket::Frontend::$Self->{Action}")->{Responsible}
             )
         {
+
+            my %ResponsibleGroupsMembers;
+            if ( !$ConfigObject->Get('Ticket::ChangeOwnerToEveryone') ) {
+                my %AllGroupsMembersNew;
+                my @QueueIDs;
+
+                if ($QueueID) {
+                    push @QueueIDs, $QueueID;
+                }
+                else {
+                    my @TicketIDs = grep {$_} $ParamObject->GetArray( Param => 'TicketID' );
+                    for my $TicketID (@TicketIDs) {
+                        my %Ticket = $TicketObject->TicketGet(
+                            TicketID      => $TicketID,
+                            DynamicFields => 0,
+                        );
+                        push @QueueIDs, $Ticket{QueueID};
+                    }
+                }
+
+                my $QueueObject = $Kernel::OM->Get('Kernel::System::Queue');
+                my $GroupObject = $Kernel::OM->Get('Kernel::System::Group');
+
+                # Put only possible 'responsible' and 'rw' agents to responsible list.
+                for my $QueueID (@QueueIDs) {
+                    my $GroupID = $QueueObject->GetQueueGroupID( QueueID => $QueueID );
+                    my %GroupMember = $GroupObject->PermissionGroupGet(
+                        GroupID => $GroupID,
+                        Type    => 'responsible',
+                    );
+                    USER_ID:
+                    for my $UserID ( sort keys %GroupMember ) {
+                        next USER_ID if !$AllGroupsMembers{$UserID};
+                        $AllGroupsMembersNew{$UserID} = $AllGroupsMembers{$UserID};
+                    }
+                    %ResponsibleGroupsMembers = %AllGroupsMembersNew;
+                }
+            }
+            else {
+                %ResponsibleGroupsMembers = %AllGroupsMembers;
+            }
+
             push @JSONData, {
                 Name         => 'ResponsibleID',
-                Data         => \%AllGroupsMembers,
+                Data         => \%ResponsibleGroupsMembers,
                 PossibleNone => 1,
             };
         }
@@ -1165,7 +1211,7 @@ sub _Mask {
             Valid => 1
         );
 
-        # only put possible rw agents to possible owner list
+        # Put only possible 'owner' and 'rw' agents to owner list.
         if ( !$ConfigObject->Get('Ticket::ChangeOwnerToEveryone') ) {
             my %AllGroupsMembersNew;
             for my $TicketID ( @{ $Param{TicketIDs} } ) {
@@ -1176,7 +1222,7 @@ sub _Mask {
                 my $GroupID = $QueueObject->GetQueueGroupID( QueueID => $Ticket{QueueID} );
                 my %GroupMember = $GroupObject->PermissionGroupGet(
                     GroupID => $GroupID,
-                    Type    => 'rw',
+                    Type    => 'owner',
                 );
                 USER_ID:
                 for my $UserID ( sort keys %GroupMember ) {
@@ -1207,7 +1253,7 @@ sub _Mask {
             Valid => 1
         );
 
-        # only put possible rw agents to possible owner list
+        # Put only possible 'responsible' and 'rw' agents to responsible list.
         if ( !$ConfigObject->Get('Ticket::ChangeOwnerToEveryone') ) {
             my %AllGroupsMembersNew;
             for my $TicketID ( @{ $Param{TicketIDs} } ) {
@@ -1218,7 +1264,7 @@ sub _Mask {
                 my $GroupID = $QueueObject->GetQueueGroupID( QueueID => $Ticket{QueueID} );
                 my %GroupMember = $GroupObject->PermissionGroupGet(
                     GroupID => $GroupID,
-                    Type    => 'rw',
+                    Type    => 'responsible',
                 );
                 USER_ID:
                 for my $UserID ( sort keys %GroupMember ) {